app/requirements.txt
aiogram==3.4.1
mysql-connector-python==8.3.0
